flask-server/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.1
pymongo==4.6.1
Werkzeug==3.0.1
transformers
torch